+++ db.json
@@ -1,0 +1,1 @@
+{"id":"abc1","title":"mytitle","author":"authorName","modifiedAt":"2021-07-24T04:00:00.000Z","publishedAt":null,"url":null,"keywords":["palabra1","palabra2","palabra3"],"readMins":20,"source":"ARTICLE"},{"id":"abc1","title":"mytitle","author":"authorName","modifiedAt":"2021-07-24T04:00:00.000Z","publishedAt":null,"url":null,"keywords":["palabra1","palabra2","palabra3"],"readMins":20,"source":"ARTICLE"},
