+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "process": "node process.js",
+    "start-basic": "node basic-server.js",
     "start": "node server.js"
   },
   "keywords": [],
@@ -24,6 +25,7 @@
   },
   "dependencies": {
     "async": "^3.2.0",
+    "express": "^4.17.1",
     "joi": "^17.4.1",
     "nodemon": "^2.0.12"
   }
